+++ routes/admin/compliance.php
@@ -5,13 +5,18 @@ declare(strict_types=1);
 use App\Domains\Compliance\Controllers\ComplianceRecordController;
 use Illuminate\Support\Facades\Route;
 
-Route::middleware(['auth'])
+Route::middleware([
+    'auth',
+    'tenant',
+    'portal.access:admin',
+])
     ->prefix('compliance')
     ->name('compliance.')
     ->group(function () {
-        Route::get('/', [ComplianceRecordController::class, 'index'])->name('index');
-        Route::post('/', [ComplianceRecordController::class, 'store'])->name('store');
-        Route::get('/{complianceRecord}', [ComplianceRecordController::class, 'show'])->name('show');
-        Route::put('/{complianceRecord}', [ComplianceRecordController::class, 'update'])->name('update');
-        Route::delete('/{complianceRecord}', [ComplianceRecordController::class, 'destroy'])->name('destroy');
+        Route::get('/', [ComplianceRecordController::class, 'index'])->middleware('permission:compliance.view')->name('index');
+        Route::post('/', [ComplianceRecordController::class, 'store'])->middleware('permission:compliance.create')->name('store');
+        Route::get('/{complianceRecord}', [ComplianceRecordController::class, 'show'])->middleware('permission:compliance.view')->name('show');
+        Route::put('/{complianceRecord}', [ComplianceRecordController::class, 'update'])->middleware('permission:compliance.update')->name('update');
+        Route::delete('/{complianceRecord}', [ComplianceRecordController::class, 'destroy'])->middleware('permission:compliance.delete')->name('destroy');
     });
+
