+++ routes/admin/projects.php
@@ -5,13 +5,18 @@ declare(strict_types=1);
 use App\Domains\Projects\Controllers\ProjectController;
 use Illuminate\Support\Facades\Route;
 
-Route::middleware(['auth'])
+Route::middleware([
+    'auth',
+    'tenant',
+    'portal.access:admin',
+])
     ->prefix('projects')
     ->name('projects.')
     ->group(function () {
-        Route::get('/', [ProjectController::class, 'index'])->name('index');
-        Route::post('/', [ProjectController::class, 'store'])->name('store');
-        Route::get('/{project}', [ProjectController::class, 'show'])->name('show');
-        Route::put('/{project}', [ProjectController::class, 'update'])->name('update');
-        Route::delete('/{project}', [ProjectController::class, 'destroy'])->name('destroy');
+        Route::get('/', [ProjectController::class, 'index'])->middleware('permission:projects.view')->name('index');
+        Route::post('/', [ProjectController::class, 'store'])->middleware('permission:projects.create')->name('store');
+        Route::get('/{project}', [ProjectController::class, 'show'])->middleware('permission:projects.view')->name('show');
+        Route::put('/{project}', [ProjectController::class, 'update'])->middleware('permission:projects.update')->name('update');
+        Route::delete('/{project}', [ProjectController::class, 'destroy'])->middleware('permission:projects.delete')->name('destroy');
     });
+
