+++ routes/admin/technician-schedules.php
@@ -5,13 +5,18 @@ declare(strict_types=1);
 use App\Domains\TechnicianSchedules\Controllers\TechnicianScheduleController;
 use Illuminate\Support\Facades\Route;
 
-Route::middleware(['auth'])
+Route::middleware([
+    'auth',
+    'tenant',
+    'portal.access:admin',
+])
     ->prefix('technician-schedules')
     ->name('technician-schedules.')
     ->group(function () {
-        Route::get('/', [TechnicianScheduleController::class, 'index'])->name('index');
-        Route::post('/', [TechnicianScheduleController::class, 'store'])->name('store');
-        Route::get('/{technicianSchedule}', [TechnicianScheduleController::class, 'show'])->name('show');
-        Route::put('/{technicianSchedule}', [TechnicianScheduleController::class, 'update'])->name('update');
-        Route::delete('/{technicianSchedule}', [TechnicianScheduleController::class, 'destroy'])->name('destroy');
+        Route::get('/', [TechnicianScheduleController::class, 'index'])->middleware('permission:technician-schedules.view')->name('index');
+        Route::post('/', [TechnicianScheduleController::class, 'store'])->middleware('permission:technician-schedules.create')->name('store');
+        Route::get('/{technicianSchedule}', [TechnicianScheduleController::class, 'show'])->middleware('permission:technician-schedules.view')->name('show');
+        Route::put('/{technicianSchedule}', [TechnicianScheduleController::class, 'update'])->middleware('permission:technician-schedules.update')->name('update');
+        Route::delete('/{technicianSchedule}', [TechnicianScheduleController::class, 'destroy'])->middleware('permission:technician-schedules.delete')->name('destroy');
     });
+
